E3_model_IA/backend/django_app/requirements-django.txt
# Django et DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
django-filter==23.3
Pillow==10.0.1
django-environ==0.11.2
drf-yasg==1.21.10

# Base de données
psycopg2-binary==2.9.7
aiosqlite

# Gestion des environnements
python-dotenv

# Tests
pytest
pytest-django
httpx

# Utilitaires
python-dateutil
requests

# Logging
loguru